remove server e2e
