+++ packages/server/lib/project-base.ts
@@ -14,7 +14,6 @@ import runEvents from './plugins/run_events'
 import Reporter from './reporter'
 import * as savedState from './saved_state'
 import { ServerCt } from './server-ct'
-import { ServerE2E } from './server-e2e'
 import { SocketCt } from './socket-ct'
 import { SocketE2E } from './socket-e2e'
 import { ensureProp } from './util/class-helpers'
@@ -23,11 +22,13 @@ import system from './util/system'
 import type { BannersState, FoundBrowser, FoundSpec, OpenProjectLaunchOptions, ReceivedCypressOptions, ResolvedConfigurationOptions, TestingType, VideoRecording } from '@packages/types'
 import { DataContext, getCtx } from '@packages/data-context'
 import { createHmac } from 'crypto'
+import { ServerBase } from './server-base'
 
 export interface Cfg extends ReceivedCypressOptions {
   projectId?: string
   projectRoot: string
   proxyServer?: Cypress.RuntimeConfigOptions['proxyUrl']
+  fileServerFolder?: Cypress.ResolvedConfigOptions['fileServerFolder']
   testingType: TestingType
   exit?: boolean
   state?: {
@@ -48,7 +49,7 @@ const debug = Debug('cypress:server:project')
 
 type StartWebsocketOptions = Pick<Cfg, 'socketIoCookie' | 'namespace' | 'screenshotsFolder' | 'report' | 'reporter' | 'reporterOptions' | 'projectRoot'>
 
-export type Server = ServerE2E | ServerCt
+export type Server = ServerBase<SocketE2E> | ServerCt
 
 export class ProjectBase<TServer extends Server> extends EE {
   // id is sha256 of projectRoot
@@ -139,7 +140,7 @@ export class ProjectBase<TServer extends Server> extends EE {
 
   createServer (testingType: Cypress.TestingType) {
     return testingType === 'e2e'
-      ? new ServerE2E() as TServer
+      ? new ServerBase() as TServer
       : new ServerCt() as TServer
   }
